move _dependencies module
