+++ services/agent/src/simcore_service_agent/core/api/rest/_dependencies.py
@@ -6,7 +6,7 @@ from typing import Annotated, cast
 from fastapi import Depends, FastAPI, Request
 from servicelib.rabbitmq._client_rpc import RabbitMQRPCClient
 
-from .settings import ApplicationSettings
+from ...settings import ApplicationSettings
 
 
 def get_application(request: Request) -> FastAPI:
